Move device code specific code to grant
